using 64 bits integers, I want avoid overflow for future extensions

+++ C/simple.c
@@ -6,21 +6,24 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
-static int finds = 0;
+
+
+static int64_t finds = 0;
 
 static void doit(void) {
-	int a, b, c;
+	int64_t a, b, c;
 
 	for (a=0; a<1000; a++) {
-		int proda = a * a * a;
-		int a10 = 1000000 * a;
+		int64_t proda = a * a * a;
+		int64_t a10 = 1000000 * a;
 		for (b=0; b<1000; b++) {
-			int prodb = b * b * b;
-			int b10 = 1000 * b;
+			int64_t prodb = b * b * b;
+			int64_t b10 = 1000 * b;
 			for (c=0; c<1000; c++) {
-				int s1 = proda + prodb + c * c * c;
-				int s2 = a10 + b10 + c;
+				int64_t s1 = proda + prodb + c * c * c;
+				int64_t s2 = a10 + b10 + c;
 				if (s1 == s2) {
 					finds++;
 					printf("(%3d, %3d, %3d => %9d) %3d^3 + %3d^3 + %3d^3 == 1000^2*%3d + 1000*%3d + %3d\n",
